Add files via upload
MazeSearch.py generates a maze and uses one of the 3 algorithms according to user's requirements. MazeSearchSpace.py is an represents the search space/every node visited in DFS search through the maze. Also attached a csv file of maze that you can take for example, and run all 3 algorithms on it.